ralph_chatbot/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
hiredis==2.3.2
prometheus-client==0.19.0
statsd==4.0.1
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
asgiref==3.7.2
uvicorn==0.27.0

# Additional dependencies for model
unsloth==2024.11.10
transformers==4.31.0
peft==0.3.0
bitsandbytes==0.41.1
xformers==0.0.20
flash-attn==1.0.9
packaging
einops
ninja
